+++ templates/base.html
@@ -30,7 +30,7 @@
     <script>
         // Apply saved settings immediately to prevent flash
         (function() {
-            const savedFont = localStorage.getItem('fontFamily') || 'system';
+            const savedFont = localStorage.getItem('fontFamily') || 'mono';
             const savedAccent = localStorage.getItem('accentColor') || '#1DA1F2';
             
             const fontMap = {
